Conform to my personal preferences
I am using the original repo as a starting point for a yeoman generator
to create more general applications. This commit removes all of the
demo react code, changes file naming conventions and directory
structures, and also fixes a few small bugs, which could be patched
back upstream. From this point forward, commits will actually be of a
managable size.